+++ hello-tango-jni-example/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name="com.google.tango.hellotangojni.MainActivity"
+            android:name="com.projecttango.hellotangonative.HelloTangoActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
