+++ hello-tango-jni-example/jni/Android.mk
@@ -13,17 +13,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-LOCAL_PATH:= $(call my-dir)
+LOCAL_PATH:= $(call my-dir)/..
 
 include $(CLEAR_VARS)
 LOCAL_MODULE := libtango-prebuilt
-LOCAL_SRC_FILES := ../../tango-service-sdk/libtango_client_api.so
+LOCAL_SRC_FILES := ../tango-service-sdk/libtango_client_api.so
 LOCAL_EXPORT_C_INCLUDES := ../tango-service-sdk/include
 include $(PREBUILT_SHARED_LIBRARY)
 
 include $(CLEAR_VARS)
 LOCAL_MODULE    := libtango-native-jni
-LOCAL_SHARED_LIBRARIES := libtango-prebuilt
-LOCAL_SRC_FILES := tango_native.c
-LOCAL_LDLIBS    := -llog
+LOCAL_SHARED_LIBRARIES := libtango-prebuilt libtangogl
+LOCAL_CFLAGS    := -Werror
+LOCAL_SRC_FILES := jni/tango_native.c
+LOCAL_C_INCLUDES := ../tango-gl-renderer ../third-party/glm/glm
+LOCAL_LDLIBS    := -llog -lGLESv2 -L$(SYSROOT)/usr/lib
 include $(BUILD_SHARED_LIBRARY)
