+++ hello-tango-jni-example/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".MainActivity" >
+    tools:context=".HelloTangoActivity" >
 
     <TextView
         android:layout_width="wrap_content"
